+++ app/Http/Livewire/Traits/WithRecommendation.php
@@ -38,6 +38,14 @@ trait WithRecommendation
         }
     }
 
+    /**
+     * 
+     */
+    public function updatedWithRecommendation()
+    {
+        $this->dispatchBrowserEvent('list-added');
+    }
+
     //* RECOMMENDS
     /**
      * 
@@ -45,7 +53,6 @@ trait WithRecommendation
     public function addRecommend()
     {
         $this->recommends[] = ['name' => '', 'jumlah_unit_needed' => "1 units"];
-        $this->dispatchBrowserEvent('list-added');
     }
     
     /**
@@ -56,14 +63,15 @@ trait WithRecommendation
      */
     public function removeRecommend($index)
     {
-        if (array_key_exists($index, $this->recommendationId)) {
-            Recommendation::where('rec_id', $this->recommendationId[$index])->delete();
+        if (array_key_exists($index, $this->recommends)) {
+            if (array_key_exists($index, $this->recommendationId)) {
+                Recommendation::where('rec_id', $this->recommendationId[$index])->delete();
+            }
+            
+            unset($this->recommends[$index]);
+            array_values($this->recommends);
+            $this->emit('unsetRecommendation');
         }
-        
-        unset($this->recommends[$index]);
-        array_values($this->recommends);
-        $this->dispatchBrowserEvent('list-added');
-        $this->emit('unsetRecommendation');
     }
 
     /**
@@ -95,6 +103,21 @@ trait WithRecommendation
         $this->setRecommendationDropdown();
     }
 
+    /**
+     * 
+     */
+    public function validateRecommends()
+    {
+        foreach ($this->recommends as $index => $recommend) {
+            $this->validate([
+                'recommends.'.$index.'.name' => 'required',
+                'recommends.'.$index.'.jumlah_unit_needed' => 'required',
+            ],[
+                'required' => 'This field is required.',
+            ]);
+        };
+    }
+
     //* MANUAL RECOMMENDS
     /**
      * 
@@ -102,7 +125,6 @@ trait WithRecommendation
     public function addManualRecommends ()
     {
         $this->manualRecommends[] = ['name' => '', 'jumlah_unit_needed' => "1 units"];
-        $this->dispatchBrowserEvent('list-added');
     }
 
     /**
@@ -112,10 +134,26 @@ trait WithRecommendation
      */
     public function removeManualRecommends($index)
     {
-        unset($this->manualRecommends[$index]);
-        array_values($this->manualRecommends);
-        $this->dispatchBrowserEvent('list-added');
-        $this->emit('unsetRecommendation');
+        if (array_key_exists($index, $this->manualRecommends)) {
+            unset($this->manualRecommends[$index]);
+            array_values($this->manualRecommends);
+            $this->emit('unsetRecommendation');
+        }
+    }
+
+    /**
+     * 
+     */
+    public function validateManualRecommends()
+    {
+        foreach ($this->manualRecommends as $index => $manualRecommend) {
+            $this->validate([
+                'manualRecommends.'.$index.'.name' => 'required',
+                'manualRecommends.'.$index.'.jumlah_unit_needed' => 'required',
+            ],[
+                'required' => 'This field is required.',
+            ]);
+        };
     }
 
     //* GENERAL RECOMMENDATIONS
